Move Event model into an own module
All ical related code is moved to an extra model module.
