+++ ng/src/gmp/models/schedule.js
@@ -20,117 +20,12 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-import moment from 'moment-timezone';
-
-import ical from 'ical.js';
-
 import {is_defined} from '../utils/identity';
 import {map} from '../utils/array';
 
-import Model from '../model.js';
+import Model from '../model';
 
-const convertIcalDate = (date, timezone) => is_defined(timezone) ?
-  moment.unix(date.toUnixTime()).tz(timezone) :
-  moment.unix(date.toUnixTime());
-
-export const ReccurenceFrequency = {
-  YEARLY: 'YEARLY',
-  MONTHLY: 'MONTHLY',
-  WEEKLY: 'WEEKLY',
-  DAILY: 'DAILY',
-  HOURLY: 'HOURLY',
-  MINUTELY: 'MINUTELY',
-  SECONDLY: 'SECONDLY',
-};
-
-class Event {
-
-  constructor(icalendar, timezone) {
-    const jcal = ical.parse(icalendar);
-    const comp = new ical.Component(jcal);
-    const vevent = comp.getFirstSubcomponent('vevent');
-    this.event = new ical.Event(vevent);
-    this.timezone = timezone;
-  }
-
-  get startDate() {
-    return convertIcalDate(this.event.startDate, this.timezone);
-  }
-
-  get duration() {
-    return this.event.duration;
-  }
-
-  get durationInSeconds() {
-    const {
-      days = 0,
-      hours = 0,
-      minutes = 0,
-      weeks = 0,
-      seconds = 0,
-    } = this.event.duration;
-    return seconds +
-      minutes * 60 +
-      hours * 60 * 60 +
-      days * 24 * 60 * 60 +
-      weeks * 7 * 24 * 60 * 60;
-  }
-
-  get recurrence() {
-    if (this.isRecurring()) {
-      const rrule = this.event.component.getFirstPropertyValue('rrule');
-      return rrule === null ? undefined : rrule;
-    }
-    return undefined;
-  }
-
-  get nextDate() {
-    if (this.isRecurring()) {
-      const now = ical.Time.now();
-      const it = this.event.iterator();
-
-      while (true) {
-        const next = it.next();
-        if (next.compare(now) >= 0) {
-          return convertIcalDate(next, this.timezone);
-        }
-      }
-    }
-    return undefined;
-  }
-
-  getNextDates(until) {
-    if (this.isRecurring()) {
-      const now = moment();
-      const it = this.event.iterator();
-      const dates = [];
-
-      while (true) {
-        const next = it.next();
-
-        if (it.completed || !next) {
-          return dates;
-        }
-
-        const mnext = convertIcalDate(next);
-
-        if (mnext.isAfter(until)) {
-          return dates;
-        }
-
-        if (mnext.isSameOrAfter(now)) {
-          dates.push(mnext);
-        }
-      }
-    }
-
-    return [];
-  }
-
-  isRecurring() {
-    return this.event.isRecurring();
-  }
-}
+import Event from './event';
 
 class Schedule extends Model {
 
@@ -142,7 +37,7 @@ class Schedule extends Model {
     const {timezone, icalendar} = elem;
 
     if (is_defined(icalendar)) {
-      ret.event = new Event(icalendar, timezone);
+      ret.event = Event.fromIcal(icalendar, timezone);
 
       delete ret.icalendar;
     }
