+++ ng/src/web/pages/schedules/component.js
@@ -30,7 +30,7 @@ import _ from 'gmp/locale.js';
 
 import {is_defined} from 'gmp/utils';
 
-import {ReccurenceFrequency} from 'gmp/models/schedule';
+import {ReccurenceFrequency} from 'gmp/models/event';
 
 import PropTypes from '../../utils/proptypes.js';
 import withGmp from '../../utils/withGmp.js';
